requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0
numpy>=1.21.0
